pyproject.toml
[tool.poetry]
name = "xycmd"
version = "0.1.0"
description = "A CLI for various services I use."
authors = ["Xyder"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
click-log = "^0.3.2"
jira = "^2.0.0"
python-dateutil = "^2.8.1"
ipdb = "^0.13.2"
toml = "^0.10.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
bandit = "^1.6.2"
pytest = "^5.3.5"
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
